requirements.txt
anyio==3.5.0
asgiref==3.5.0
astroid==2.9.3
attrs==21.4.0
cfgv==3.3.1
click==8.0.4
distlib==0.3.4
Faker==13.3.2
fastapi==0.75.0
filelock==3.6.0
h11==0.13.0
identify==2.4.11
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
nodeenv==1.6.0
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.17.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pydantic==1.9.0
pylint==2.12.2
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.19.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
Telethon==1.24.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
uvicorn==0.17.5
virtualenv==20.13.3
Werkzeug==2.0.3
wrapt==1.13.3
